Rename Editor src files to avoid exporting
